refactor: user detail page structure
